move main logic to cli.h
